AMP-13389
git-svn-id: file:///svn/branches/AMP_2_1_RELEASE@18521 9e8d75aa-d0cb-44a9-8341-a4b8ba1a38b4

+++ amp/WEB-INF/src/org/digijava/module/gis/util/DbUtil.java
@@ -2241,7 +2241,7 @@ public class DbUtil {
                     Object[] resArray = (Object[]) resultObj;
                     Long actId = (Long)resArray[0];
                     Long prgId = (Long)resArray[1];
-                    Float prgPercentage = new Float(((Long)resArray[2]).floatValue());
+                    Float prgPercentage = new Float((Float) resArray[2]);
 
                     if (retVal.containsKey(actId)) {
                         ((Map<Long, Float>) retVal.get(actId)).put(prgId, prgPercentage);
